+++ R/exclude_pars.R
@@ -19,7 +19,7 @@ exclude_pars.brmsfit <- function(x, save_ranef = TRUE, save_mevars = FALSE,
   save_mevars <- as_one_logical(save_mevars)
   save_all_pars <- as_one_logical(save_all_pars)
   out <- character(0)
-  bterms <- parse_bf(x$formula)
+  bterms <- brmsterms(x$formula)
   c(out) <- exclude_pars(
     bterms, data = x$data, save_ranef = save_ranef, 
     save_all_pars = save_all_pars, save_mevars = save_mevars, ...