+++ R/families.R
@@ -1367,7 +1367,7 @@ summarise_links.mvbrmsformula <- function(x, wsp = 0, ...) {
 
 #' @export
 summarise_links.brmsformula <- function(x, mv = FALSE, ...) {
-  x <- parse_bf(x)
+  x <- brmsterms(x)
   dpars <- valid_dpars(x)
   links <- setNames(rep("identity", length(dpars)), dpars)
   links_pred <- ulapply(x$dpars, function(x) x$family$link)